requirements.txt
Pillow==10.4.0
ttkthemes==3.2.2
PyPDF2==3.0.1